requirements.txt
openai-whisper
transformers
torch
scikit-learn
numpy